[BAQE-829] Create separate test class for each client
